+++ page/common.js
