+++ page/home.js
