+++ server/home.js
@@ -1,6 +1,6 @@
 import {serveStaticFile} from './file.js'
 import {readLatestDevices} from '../db/stats.js'
-import {createHome} from '../pages/home.js'
+import {createHome} from '../page/home.js'
 
 
 export default async function setup(app) {
